Uploaded prototype 1 month early
This may or may not be the final build, but here's something for you to use in the next month until proper release. Expect the files in this commit to POSSIBLY be obsolete then. If it isn't, expect me to rename them. I also need to make a Github Pages for this.