PC98 version (ANEX86.BMP replacement)
I used NP21 Wii to make this, using the TrueType version. This is the same process I used to make the DOSBox-X FREECG98.BMP I did in 2018 using an ancestor of UnifontEX. Of course, nowadays DOSBox-X supports TrueType loading directly. Also, if you *are* looking for an embedded systems font sheet, I strongly advise using the TTF2PNG version instead. It's unabridged, and it's easier to read by computers (because it can be read sequentially). Just be sure not to anger the FSF. Otherwise, I don't particularly care what people do with it (or quite frankly any of my creations), so long as it isn't disgusting or bigoted. Now I have to amend the readme yay.