Added extra tables
So FontForge's newest portable version does not break when you tell it to add JSTF, BASE, and MATH at the same time. Also CacheTT will no longer break Windows thinking it is OpenType because FontForge makes it so only the VDMX can be generated by it, and VDMX alone does not trigger the bug. So renderer compatibility should be better. Also no tables get erroneously removed, and Winamp still loads the font, even after going through CacheTT. And yes, I had set vertical baselines in FontForge, so in combination with the VDMX table, vertical text should load better (apparently HDMX and LTSH aren't generated, but seeing as prototypes that had THOSE would suffer the bug of Windows not showing them as OpenType, I feel like that isn't a bad thing per se. Also VDMX is only in the TTF and EOT versions, but BASE and MATH aren't limited to those. Keep in mind that this pushes the WOFF1 version over 3 and 1/3 MB in decimal notation, but only by less than 100 bytes, so it's not a problem. Having better math support and better baselines is an upside. Oh and CacheTT does not work on WOFFs, so I would have to convert the TrueType into a WOFF+WOFF2 externally. TTF2WOFF is used to improve FontForge's WOFF1 compression via Zopfli, but it doesn't have the ability to add the relevant version number when making a WOFF1 out of a TTF. It just would get ugly, and not having a version number could break compatibility. At least VDMX is valid to have in a webfont, but the table is 1504 bytes, and it is values generated from sizes 1 to 256. Also this does not change the fact that LTSH, HDMX, and VDMX can be hated by picky renderers, but as established earlier, VDMX alone is well-behaved, but even so, I don't want to risk buggy browsers.)