VDMX table on WOFF1
Oh and fans of Mew will approve of the last 80 bytes of the file. I got "lucky" that the WOFF was 3,334,256 bytes, but I decided to use the Private Data feature of WOFF that allows arbitrary binary data to throw in a special Easter egg. I originally planned to use my 3081 byte demo, but because it ALSO is DEFLATEd, I worried that it could interfere with decoding. Plus, 3081 bytes of extra overhead for an Easter egg on a webfont is a bit much. HOWEVER, the boxed PK3 file for the Shiny 6x31IV RNG Mew of mine from 2017 at 80 bytes is FAR more manageable, and actually aligned, so it shouldn't mess up anything. I don't know if the full 100 byte PK3 Party would be safe, but 80 IS a multiple of 16 so it has that going for it. Plus it is MUCH easier to extract, and the total WOFF file size is a visually-prettier number, 3,334,336 bytes. Remember, this is all to enable the helpful-for-vertical-text VDMX table, to assist renderers. Since UnifontEX as a project is designed around making use of features (and even formats, among other decisions) never used by upstream Unifont, it's only fitting for when we're using WOFF(2), a format never used by Unifont, to make use of exclusive features in it. The full XML metadata format has no easily-available generator, and while it IS compressed, it's honestly redundant given how I filled out all the TrueType fields, AND it doesn't have as many features there. It's better not to play THAT game. However 80 bytes of Private Data is a much healthier way of getting mileage out of the WOFF(2)-exclusive features in keeping with UnifontEX's use of things never originally done by upstream Unifont. And what better data for a nonbinary otherkin to use for this then an OP PK3 file of theirs of their favorite mythical, that just so happens to comprise their sona? If it makes the size value more orderly, then so be it. And this build is intended for making vertical text (and anything else affected by VDMX) look better. Oh and hdmx and LTSH actually reduce the effectiveness of accessibility scaling on Android 14, and accessibility is why there's no hinting (hinting can cause strobing in animated text). Basically, UnifontEX is designed for high compatibility, plus accessibility, and using features never originally used by the original designers. I use it as my Android UI font.