Add memorial message to one of the muses of Unifont's inventor
It's been 30 years since the original attempts by Roman Cyzborra to make Unifont. Using UnifontEX metadata fields for a rolling obituary does not a final build make (pardon my Yoda.)

You can't consider a font stable enough to use in a UI if it's being a rolling obituary or an endlessly-updating copyright notice. Oh and I haven't made a new Github release based on the current repo state with the VDMX WOFF1, because I'm attempting to put the VDMX table and Easter egg in the WOFF2 version too, but so far the needed tools to achieve the desired results are proving elusive but this isn't a wild goose chase so I'm reluctant to rush it out. I should also mention that the professor was mentioned to have played specifically the role of a thesis advisor to Roman, taught computer science (not linguistics), and this all took place in the early 1990s where digital typography and especially Unicode were quite young. Combined with the fact that Roman Cyzborra invented his own glyph format unlike all others, and the fact that BWTC32Key, UnifontEX, and my other stuff isn't fully understood by quite a few of my nerdiest professors even now (25-30 years after Roman's time, and yes I'm in college for the record) makes me suspect that glyph drawing may not have been this professor's role in the project. Professors are very busy people too, so drawing glyphs 1-on-1 with a student for ages in a custom format would be less likely. I'm leaving the credit to Mark Crispin (the professor hadn't been given anything but a surname on Roman's Unifont page, making crediting difficult especially given that I didn't know what university or region on his page about Unicode itself, about an hour ago I found his name) in because if he was a friend of Roman's, he may have potentially even worked together with Roman on some glyphs. I myself bounce quite a lot of creative decisions off of my friends, online and offline, sometimes even having friends sit next to me as I do my thing in my spare time, or sending WIPs out on Discord. Either way I still felt it was necessary to issue the memorial to one of Roman's muses since Unifoundry hasn't.

 Also the professor being from a Berlin university serves to explain why upstream Unifont is registered in Germany. Turns out it IS because Roman in Germany did Unifont, not to deliberately dodge America's lack of a copyright protection for bitmap fonts (obviously not TrueType ones.) That factor may not apply to upstream Unifont, but *I* am American, and I live in the USA at present, and was born in the USA, and I made UnifontEX in the United States. In the United States, American derivatives of foreign works are bound by United States law. Meaning that UnifontEX in something like TTF2PNG may in fact lack copyrightability in a way that upstream Unifont won't. If you ARE going to rely on that in your use case, the TTF2PNG version would certainly constitute the safest version for that situation. I still can't recommend angering the FSF in your use case though, as much as I want to see UnifontEX as a UI font to banish most mojibake. But if the TTF2PNG version is in fact public domain, then legally I cannot stop you from integrating it into your use cases. That being said, Android is open-source, Darwin and WSLg+WSA exist, and so this means that even the big OS vendors have potential ability to use even the TrueType version lol. Also then there's the  Kindle OS using Linux and OpenDyslexic, as well as Fanfiction.net using OpenDyslexic, so in theory stopping this being used as a UI font may not be simple or easy, in fact it may be quite the opposite, and even messy. I hate messy situations.