Uploaded example CSS file
This is an @font-face CSS file that utilizes the fact that I've made TTF, EOT, SVG, WOFF, and WOFF2 builds, AKA: ALL the webfont formats. Pop this CSS file on your server, (or edit it to have the font blobs on your own server) to upgrade Unicode support across browsers. I'm going to put the CSS file in the main directory as UnifontExMono.css