Make HTTPS-enforcing browsers load VDMX tables
Since the WOFF2 is inferior due to lacking VDMX tables (I can't find an app that turns TTF into WOFF2 with the arbitrary data function), I've made it so that it's loaded over http to make modern Gecko and modern Blink not load it, because they still support WOFF1, also WOFF1 compresses the best, and has an easter egg. TTF apparently gets used only if WOFFs fail to download. Basically, I've made it so you get the Easter egg and better compression, but if not, then you get to still keep the VDMX table. Gotta have good vertical text!