Add note about 15.1 rebasing being impossible, and mitigations too
I spent the day trying to see if I could make a build with characters from 15.1.01. It didn't work when it got to the webfont stage after several tries and edits. Somehow on February 2nd, 2024, I got VERY lucky, because of multiple 1/16 odds stacking together to make the TrueType AND WOFF1 multiples of 16 bytes. Importantly, aligned in the right way too. I'm lucky I got the VDMX table in. One more character in the comments or FONTLOG would have been disastrous here.

I got VERY lucky. Also it's simpler (in a GOOD way).