PNG+note
TTF2PNG + GIMP on this particular image did *not* get smaller like how the jump from Unifont 11 to 15 did. It actually got bigger after the 5 new characters. To make it fit 1MiB I used ECT and Leanify on it, though I DID ask ECT to keep filter mode zero and 1bpp. Sadly, the IDAT chunks got grouped. So decoding it low-level can't rely on the old chunk system. But going up a Unicode version is a good thing. As I've mentioned before, this update will take a bit of getting used to.