Add retro-safe version
OK, so it turns out after going further and using Defluff, Deflopt, and OxiPNG it's possible to hit 946KiB, and through heavily modifying jspngopt, I was able to make it forcibly revert to the old 8192-byte chunks of the Unicode 15.0 version AND there is not over 128 of them. The old version had 126, this version has 128 on the dot. It will thus still work on retro systems. AND it's under 1MiB. Now even the casualty that was IDAT merging has been undone. AND I didn't break out of 7-bit chunk count. Time to re-issue everything!